# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rave privado V1",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v1/account": {
      "patch": {
        "operationId": "PATCH__api_v1_account",
        "summary": "PATCH /api/v1/account",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Origin",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Origen exacto de Rave"
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación completada"
          },
          "400": {
            "description": "Payload inválido"
          },
          "401": {
            "description": "Sesión requerida"
          },
          "403": {
            "description": "Permiso, Origin, CSRF o reautenticación requerido"
          },
          "409": {
            "description": "Conflicto de estado"
          },
          "429": {
            "description": "Límite excedido"
          }
        },
        "x-rave-access": "user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "displayName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 60
                  },
                  "preferences": {
                    "type": "object",
                    "properties": {
                      "volume": {
                        "type": "number",
                        "minimum": 0,
                        "maximum": 1
                      },
                      "muted": {
                        "type": "boolean"
                      },
                      "subtitleId": {
                        "anyOf": [
                          {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "subtitleLanguage": {
                        "anyOf": [
                          {
                            "type": "string",
                            "maxLength": 80
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "subtitleOffset": {
                        "type": "number",
                        "minimum": -5,
                        "maximum": 5,
                        "multipleOf": 0.25
                      },
                      "subtitleSize": {
                        "type": "number",
                        "minimum": 75,
                        "maximum": 200
                      },
                      "subtitleBackground": {
                        "type": "boolean"
                      },
                      "quality": {
                        "type": "string",
                        "maxLength": 80
                      }
                    },
                    "additionalProperties": false
                  },
                  "avatarAssetId": {
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                },
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/uploads": {
      "post": {
        "operationId": "POST__api_v1_admin_uploads",
        "summary": "POST /api/v1/admin/uploads",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Origin",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Origen exacto de Rave"
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Misma clave/payload para repetir una intención; no reutilizar con otro payload"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación completada"
          },
          "400": {
            "description": "Payload inválido"
          },
          "401": {
            "description": "Sesión requerida"
          },
          "403": {
            "description": "Permiso, Origin, CSRF o reautenticación requerido"
          },
          "409": {
            "description": "Conflicto de estado"
          },
          "429": {
            "description": "Límite excedido"
          }
        },
        "x-rave-access": "owner",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "description": {
                    "default": "",
                    "type": "string",
                    "maxLength": 4000
                  },
                  "category": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 80
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "originalName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "expectedBytes": {
                    "type": "integer",
                    "exclusiveMinimum": 0,
                    "maximum": 21474836480
                  }
                },
                "required": [
                  "title",
                  "description",
                  "originalName",
                  "expectedBytes"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/uploads/{id}": {
      "head": {
        "operationId": "HEAD__api_v1_admin_uploads__id",
        "summary": "HEAD /api/v1/admin/uploads/:id",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Offset durable; sin cuerpo",
            "headers": {
              "Upload-Offset": {
                "schema": {
                  "type": "integer"
                }
              },
              "Upload-Length": {
                "schema": {
                  "type": "integer"
                }
              },
              "Upload-State": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Payload inválido"
          },
          "401": {
            "description": "Sesión requerida"
          },
          "403": {
            "description": "Permiso, Origin, CSRF o reautenticación requerido"
          },
          "409": {
            "description": "Conflicto de estado"
          },
          "429": {
            "description": "Límite excedido"
          }
        },
        "x-rave-access": "owner"
      },
      "patch": {
        "operationId": "PATCH__api_v1_admin_uploads__id",
        "summary": "PATCH /api/v1/admin/uploads/:id",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Origin",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Origen exacto de Rave"
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Upload-Offset",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "in": "header",
            "name": "Content-Length",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Offset durable; sin cuerpo",
            "headers": {
              "Upload-Offset": {
                "schema": {
                  "type": "integer"
                }
              },
              "Upload-Length": {
                "schema": {
                  "type": "integer"
                }
              },
              "Upload-State": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Payload inválido"
          },
          "401": {
            "description": "Sesión requerida"
          },
          "403": {
            "description": "Permiso, Origin, CSRF o reautenticación requerido"
          },
          "409": {
            "description": "Conflicto de estado"
          },
          "429": {
            "description": "Límite excedido"
          }
        },
        "x-rave-access": "owner",
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "DELETE__api_v1_admin_uploads__id",
        "summary": "DELETE /api/v1/admin/uploads/:id",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Origin",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Origen exacto de Rave"
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Misma clave/payload para repetir una intención; no reutilizar con otro payload"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación completada"
          },
          "400": {
            "description": "Payload inválido"
          },
          "401": {
            "description": "Sesión requerida"
          },
          "403": {
            "description": "Permiso, Origin, CSRF o reautenticación requerido"
          },
          "409": {
            "description": "Conflicto de estado"
          },
          "429": {
            "description": "Límite excedido"
          }
        },
        "x-rave-access": "owner",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {},
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
    "/api/v1/media/{id}": {
      "get": {
        "operationId": "GET__api_v1_media__id",
        "summary": "GET /api/v1/media/:id",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación completada"
          },
          "400": {
            "description": "Payload inválido"
          },
          "401": {
            "description": "Sesión requerida"
          },
          "403": {
            "description": "Permiso, Origin, CSRF o reautenticación requerido"
          },
          "409": {
            "description": "Conflicto de estado"
          },
          "429": {
            "description": "Límite excedido"
          }
        },
        "x-rave-access": "user"
      }
    },
...
    "/api/v1/account/avatar": {
      "post": {
        "operationId": "POST__api_v1_account_avatar",
        "summary": "POST /api/v1/account/avatar",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Origin",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Origen exacto de Rave"
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Misma clave/payload para repetir una intención; no reutilizar con otro payload"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación completada"
          },
          "400": {
            "description": "Payload inválido"
          },
          "401": {
            "description": "Sesión requerida"
          },
          "403": {
            "description": "Permiso, Origin, CSRF o reautenticación requerido"
          },
          "409": {
            "description": "Conflicto de estado"
          },
          "429": {
            "description": "Límite excedido"
          }
        },
        "x-rave-access": "user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "imageBase64": {
                    "type": "string",
                    "maxLength": 3145728
                  }
                },
                "required": ["imageBase64"],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/drive/import": {
      "post": {
        "operationId": "POST__api_v1_admin_drive_import",
        "summary": "POST /api/v1/admin/drive/import",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Origin",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Origen exacto de Rave"
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Misma clave/payload para repetir una intención; no reutilizar con otro payload"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación completada"
          },
          "400": {
            "description": "Payload inválido"
          },
          "401": {
            "description": "Sesión requerida"
          },
          "403": {
            "description": "Permiso, Origin, CSRF o reautenticación requerido"
          },
          "409": {
            "description": "Conflicto de estado"
          },
          "429": {
            "description": "Límite excedido"
          }
        },
        "x-rave-access": "owner",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "fileIds": {
                    "minItems": 1,
                    "maxItems": 20,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "pattern": "^[A-Za-z0-9_-]{10,200}$"
                    }
                  }
                },
                "required": ["fileIds"],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "rave"
      }
    }
  }
}